+++ core/src/main/java/nl/nn/adapterframework/configuration/IbisContext.java
@@ -73,7 +73,6 @@ public class IbisContext {
 	private static final String INSTANCE_NAME = APP_CONSTANTS.getResolvedProperty("instance.name");
 	private static final String CONFIGURATIONS = APP_CONSTANTS.getResolvedProperty("configurations.names.application");
 	private static final String APPLICATION_SERVER_TYPE_PROPERTY = "application.server.type";
-	private static final String FLOW_CREATE_DIAGRAM_URL = APP_CONSTANTS.getResolvedProperty("flow.create.url");
 	private static final long UPTIME = System.currentTimeMillis();
 
 	static {
@@ -177,10 +176,6 @@ public class IbisContext {
 			MessageKeeper messageKeeper = new MessageKeeper();
 			messageKeepers.put("*ALL*", messageKeeper);
 
-			if (StringUtils.isNotEmpty(FLOW_CREATE_DIAGRAM_URL)) {
-				flowDiagram = new FlowDiagram(FLOW_CREATE_DIAGRAM_URL);
-			}
-
 			applicationContext = createApplicationContext();
 			ibisManager = (IbisManager)applicationContext.getBean("ibisManager");
 			ibisManager.setIbisContext(this);
@@ -190,6 +185,11 @@ public class IbisContext {
 			registerApplicationModules();
 			load();
 			getMessageKeeper().setMaxSize(Math.max(messageKeeperSize, getMessageKeeper().size()));
+
+			//TODO consider moving this to #FlowDiagram
+			String graphvizJsVersion = APP_CONSTANTS.getProperty("graphviz.js.version", null);
+			String graphvizJsFormat = APP_CONSTANTS.getProperty("graphviz.js.format", null);
+			flowDiagram = new FlowDiagram(graphvizJsFormat, graphvizJsVersion);
 
 			log("startup in " + (System.currentTimeMillis() - start) + " ms");
 		}
@@ -549,13 +549,11 @@ public class IbisContext {
 
 	private void generateFlow() {
 		if (flowDiagram != null) {
-			List<Configuration> configurations = ibisManager
-					.getConfigurations();
+			List<Configuration> configurations = ibisManager.getConfigurations();
 			try {
 				flowDiagram.generate(configurations);
 			} catch (Exception e) {
-				log("*ALL*", null, "error generating flowDiagram",
-						MessageKeeperMessage.WARN_LEVEL, e);
+				log("*ALL*", null, "error generating flowDiagram", MessageKeeperMessage.WARN_LEVEL, e);
 			}
 		}
 	}