+++ RELEASES.md
@@ -16,6 +16,8 @@ Ibis AdapterFramework release notes
 - Use XSLT 2.0 instead of 1.0 for configuration tweaks (e.g. stub4testtool.xsl)
 - Display showConfig jdbc errors per realm
 - Add heath endpoints for individual adapters and entire application
+- Add Google's V8 JavaScript engine
+- Use VizJs to generate flow diagrams